+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # CHANGELOG
 
 ## Unreleased
+**Changes**
+* [Breaking] `platformPayParams` has changed from `PlatformPay.PaymentMethodParams` to `Onramp.OnrampPlatformPayParams` to better match expectations for GooglePay. ([#2390] (https://github.com/stripe/stripe-react-native/pull/2390))
+* [Added] Now accepts `isEmailRequired` and `allowCreditCards` as part of `GooglePayConfig`. ([#2390] (https://github.com/stripe/stripe-react-native/pull/2390))
 
 **Fixes**
 * [Fixed] `autocompleteCountries` prop on `AddressSheet` was silently ignored on iOS.